+++ CHANGELOG.md
@@ -21,7 +21,7 @@
 - Updated the getID3 library to the release version 1.9.24-202509040923
 - OpenSubsonic API: 
   * Add property `mediaType` to the result type `Child`
-  * Add property `genres` to the album results
+  * Add properties `genres` and `artists` to the album results
 
 ### Fixed
 - Playback stopping when clicking the title "0 results" of an empty advanced search result